Feat: 사이드바 북마크 데이터 불러오는 위치 변경
